+++ Changelog.markdown
@@ -4,6 +4,7 @@
 
 *Unreleased*
 
+* Add `rowBackgroundColor` to SSCollectionView. This is useful for having a texture background color on the collection view and preventing the internal UITableView from using it as the background color of each row.
 * Add `NSAssert` in SSCollectionView for a `nil` `reuseIdentifier`. This used to be an `NSLog`. It was changed to an assert because you really need to provide one or your performance will be just awful. Be user to add `NS_BLOCK_ASSERTIONS` in your release build to avoid crashing just incase you missed one somewhere.
 * Add `deepMutableCopy` to `NSDictionary` and `NSArray`